+++ src/controllers/radioChannelController.js
@@ -1,4 +1,17 @@
-import { findRadioChannelById } from "../services/radioChannelService.js";
+import {
+  findRadioChannelList,
+  findRadioChannelById,
+} from "../services/radioChannelService.js";
+
+export const getRadioChannelList = async (_req, res, next) => {
+  try {
+    const channels = await findRadioChannelList();
+
+    return res.status(200).json(channels);
+  } catch (error) {
+    next(error);
+  }
+};
 
 export const getRadioChannelById = async (req, res, next) => {
   try {
