--- src/routes/radio-channels/allRadioList.js
@@ -1,9 +1,0 @@
-import express from "express";
-
-import { getRadioChannelList } from "../../controllers/radioController.js";
-
-const router = express.Router();
-
-router.get("", getRadioChannelList);
-
-export default router;
