+++ src/routes/radio-channels/index.js
@@ -1,10 +1,14 @@
 import express from "express";
 
-import { getRadioChannelById } from "../../controllers/radioChannelController.js";
+import {
+  getRadioChannelList,
+  getRadioChannelById,
+} from "../../controllers/radioChannelController.js";
 import { validateChannelId } from "../../validators/channelValidator.js";
 
 const router = express.Router();
 
+router.get("", getRadioChannelList);
 router.get("/:channelId", validateChannelId, getRadioChannelById);
 
 export default router;
