+++ src/server.js
@@ -3,8 +3,7 @@ import express from "express";
 
 import { errorHandler } from "./middlewares/errorHandler.js";
 import adKeywordRoutes from "./routes/ad-keywords/index.js";
-import radioChannelRoutes from "./routes/radio-channels/[channelId].js";
-import radioRoutes from "./routes/radio-channels/allRadioList.js";
+import radioChannelRoutes from "./routes/radio-channels/index.js";
 import reportRouter from "./routes/report/index.js";
 import userRoutes from "./routes/users/index.js";
 
@@ -14,11 +13,10 @@ const app = express();
 
 app.use(express.json());
 
+app.use("/ad-keywords", adKeywordRoutes);
+app.use("/radio-channels", radioChannelRoutes);
 app.use("/reports", reportRouter);
 app.use("/users", userRoutes);
-app.use("/radio-channels", radioChannelRoutes);
-app.use("/radio-channels", radioRoutes);
-app.use("/ad-keywords", adKeywordRoutes);
 
 app.use(errorHandler);
